+++ samplesheets/tests/test_views_api_taskflow.py
@@ -530,12 +530,12 @@ class TestIrodsAccessTicketUpdateAPIView(IrodsAccessTicketAPIViewTestBase):
             'sodar_uuid': str(self.ticket.sodar_uuid),
             'label': LABEL_UPDATE,
             'ticket': self.ticket.ticket,
-            'assay': self.ticket.assay.pk,
-            'study': self.ticket.study.pk,
+            'study': str(self.study.sodar_uuid),
+            'assay': str(self.assay.sodar_uuid),
             'path': self.ticket.path,  # Path should not be updated
             'date_created': local_date_created.isoformat(),
             'date_expires': self.date_expires_update,
-            'user': self.ticket.user.pk,
+            'user': self.get_serialized_user(self.user),
             'is_active': self.ticket.is_active(),
         }
         self.assertEqual(response.json(), expected)
@@ -600,12 +600,12 @@ class TestIrodsAccessTicketUpdateAPIView(IrodsAccessTicketAPIViewTestBase):
             'sodar_uuid': str(self.ticket.sodar_uuid),
             'label': TICKET_LABEL,
             'ticket': self.ticket.ticket,
-            'assay': self.ticket.assay.pk,
-            'study': self.ticket.study.pk,
+            'study': str(self.study.sodar_uuid),
+            'assay': str(self.assay.sodar_uuid),
             'path': self.ticket.path,
             'date_created': local_date_created.isoformat(),
             'date_expires': self.date_expires_update,
-            'user': self.ticket.user.pk,
+            'user': self.get_serialized_user(self.user),
             'is_active': self.ticket.is_active(),
         }
         self.assertEqual(response.json(), expected)